requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
boto3==1.34.139
    # via -r requirements.in
botocore==1.34.139
    # via
    #   boto3
    #   s3transfer
certifi==2024.7.4
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via -r requirements.in
cryptography==42.0.8
    # via -r requirements.in
idna==3.7
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
oauthlib==3.2.2
    # via
    #   -r requirements.in
    #   requests-oauthlib
pycparser==2.22
    # via cffi
pygments==2.18.0
    # via rich
pyjwt==2.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
requests==2.32.3
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via -r requirements.in
rich==13.7.1
    # via -r requirements.in
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==2.2.2
    # via
    #   botocore
    #   requests
